[IMPROVE] realtime emoji animation (#118)
* improve: realtime emoji animation

* working with prod socket conn and removing env deps
