+++ app/.env.local.sample
@@ -36,13 +36,5 @@ ROCKETCHAT_CLIENT_SECRET="client secret"
 ROCKETCHAT_URL="base url of rocket chat server"
 ### END: Environment variables for RocketChat OAuth legacy
 
-# INAPPCHAT
-# for development it defaults to: http://localhost:3000
-# in prod, it should be: [PROTOCOL]://[DOMAIN]:[PORT]
-NEXT_PUBLIC_ROCKET_CHAT_HOST=requiredInProd
-# id of the room where conf will take place
-NEXT_PUBLIC_ROCKET_CHAT_CONF_RID=required
-# id of the greenroom
-NEXT_PUBLIC_ROCKET_CHAT_GREENROOM_RID=required
 NEXT_PUBLIC_ROCKET_CHAT_GREENROOM_RTMP="rtmp://bkk.contribute.live-video.net/app/{stream_key}"
